rt(d3d12): move update subresources to lut
